Move program to actions
